requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
dj-database-url==3.0.0
Django==5.2.1
django-appconf==1.1.0
django-ckeditor==6.3.2
django-imagekit==5.0.0
django-js-asset==3.1.2
django-recaptcha==4.1.0
django-user-agents==0.4.0
frozenlist==1.6.0
gunicorn==23.0.0
idna==3.10
instamojo-wrapper==1.2.0
multidict==6.4.4
mysqlclient==2.2.7
openai==0.28.0
packaging==25.0
paypalrestsdk==1.5.0
pilkit==3.0
pillow==11.2.1
propcache==0.3.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
razorpay==1.4.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
sslcommerz-lib==1.0
stripe==12.2.0
tqdm==4.67.1
twilio==9.6.2
typing_extensions==4.13.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
Unidecode==1.4.0
urllib3==1.25.11
user-agents==2.2.0
whitenoise==6.5.0
yarl==1.20.0
